cache the toplevel channel for each channel.
this simplifies checking if a channel belongs to a given store and the
creation and maintenance of sub-channels by allowing stores to only define
which top level channels they include. Then all subchannels from there on
down will reflect the topLevel and can be checked for store inclusion.

this implies that a top-level channel is an all-or-nothing thing: a store
that includes a given top level item includes ALL of its subchannels too.
however, i could not think of a rational use case where this would be necessary
or even desired.

(in fact: i ran into the problem of subchannels not associated with a store
because only parent channels had been while writing unit tests ...)