make update runs smoother
previously it was shift()ing the array passed in, leading to the first
item in the array being chopped which meant the first update was never
stored in updates/completed.

shift() was also returning an array object. needed to toString() it so
that we would have a string and a reliable check for .sql/.plsql endings