Fix the utils getUrl code
I specifically didn't want it to accept anything because of the
way our server works right now. One of the most common ways in
which paths are messed up leads to returns of html instead of json,
and without any warnings the later code would just weirdly fail
because getUrl would accept anything.