Merge the incoming and published info.
When posting assets we need to know if the asset has already
been published and the posting is simply to update some of its
info, or whether it's the initial posting of a brand new asset.
The latter case comes with lots of requirements (tags, previews
and filled fields in the assets table), the latter just lets
everything through (e.g. just the description could be set).