Implement rejection of assets.
We can now reject posted assets. Rejection reason is required
because it's used in the email sent to the publisher explaining
the reasons for the rejection. Rejection doesn't delete an
asset, just remove the 'posted' flag so that the vetters concerns
can be addressed without having to recreate and reupload all
the data.