Try to fix the wrap transaction function.
It's really evil. Split up the one that tries to send automatic
replies from the one that doesn't. The one that was using callbacks
was broken - it was sending a json reply if the second argument
existed which it always did for the async version but it was a db
so it would try to serialize the db objects and send it as a json
reply...