this search *actually* works
with the full text searching, "appr" would not match "approval". obviously
not what we want. full text search is for taking a full word like "Hummingbird"
and finding it and its variants in text. it does not do random substring
matching which is what we need here.

using ~ is a bit slower (~50% with the ~42k tags on the production server) but
that turns out out to be a difference of 4ms